+++ src/Unity.Mathematics/Il2CppEagerStaticClassConstructionAttribute.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Unity.IL2CPP.CompilerServices
+{
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, Inherited = false, AllowMultiple = false)]
+    public class Il2CppEagerStaticClassConstructionAttribute : Attribute
+    {
+    }
+}