+++ src/Unity.Mathematics/double4.gen.cs
@@ -9,6 +9,7 @@
 using System;
 using System.Runtime.CompilerServices;
 using System.Diagnostics;
+using Unity.IL2CPP.CompilerServices;
 
 #pragma warning disable 0660, 0661
 
@@ -16,6 +17,7 @@ namespace Unity.Mathematics
 {
     [DebuggerTypeProxy(typeof(double4.DebuggerProxy))]
     [System.Serializable]
+    [Il2CppEagerStaticClassConstruction]
     public partial struct double4 : System.IEquatable<double4>, IFormattable
     {
         public double x;