+++ src/Unity.Mathematics/math.cs
@@ -1,9 +1,11 @@
 using System;
 using System.Runtime.InteropServices;
 using System.Runtime.CompilerServices;
+using Unity.IL2CPP.CompilerServices;
 
 namespace Unity.Mathematics
 {
+    [Il2CppEagerStaticClassConstruction]
     public static partial class math
     {
         /// <summary>Extrinsic rotation order. Specifies in which order rotations around the principal axes (x, y and z) are to be applied.</summary>
@@ -447,8 +449,8 @@ namespace Unity.Mathematics
         /// Checks if each component of the input is a power of two.
         /// </summary>
         /// <remarks>If a component of x is less than or equal to zero, then this function returns false in that component.</remarks>
-        /// <param name="x"><see cref="int2"/> input</param>
-        /// <returns><see cref="bool2"/> where true in a component indicates the same component in the input was a power of two.</returns>
+        /// <param name="x"><see cref="Mathematics.int2"/> input</param>
+        /// <returns><see cref="Mathematics.bool2"/> where true in a component indicates the same component in the input was a power of two.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool2 ispow2(int2 x)
         {
@@ -459,8 +461,8 @@ namespace Unity.Mathematics
         /// Checks if each component of the input is a power of two.
         /// </summary>
         /// <remarks>If a component of x is less than or equal to zero, then this function returns false in that component.</remarks>
-        /// <param name="x"><see cref="int3"/> input</param>
-        /// <returns><see cref="bool3"/> where true in a component indicates the same component in the input was a power of two.</returns>
+        /// <param name="x"><see cref="Mathematics.int3"/> input</param>
+        /// <returns><see cref="Mathematics.bool3"/> where true in a component indicates the same component in the input was a power of two.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool3 ispow2(int3 x)
         {
@@ -471,8 +473,8 @@ namespace Unity.Mathematics
         /// Checks if each component of the input is a power of two.
         /// </summary>
         /// <remarks>If a component of x is less than or equal to zero, then this function returns false in that component.</remarks>
-        /// <param name="x"><see cref="int4"/> input</param>
-        /// <returns><see cref="bool4"/> where true in a component indicates the same component in the input was a power of two.</returns>
+        /// <param name="x"><see cref="Mathematics.int4"/> input</param>
+        /// <returns><see cref="Mathematics.bool4"/> where true in a component indicates the same component in the input was a power of two.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool4 ispow2(int4 x)
         {
@@ -495,8 +497,8 @@ namespace Unity.Mathematics
         /// Checks if each component of the input is a power of two.
         /// </summary>
         /// <remarks>If a component of x is less than or equal to zero, then this function returns false in that component.</remarks>
-        /// <param name="x"><see cref="uint2"/> input</param>
-        /// <returns><see cref="bool2"/> where true in a component indicates the same component in the input was a power of two.</returns>
+        /// <param name="x"><see cref="Mathematics.uint2"/> input</param>
+        /// <returns><see cref="Mathematics.bool2"/> where true in a component indicates the same component in the input was a power of two.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool2 ispow2(uint2 x)
         {
@@ -507,8 +509,8 @@ namespace Unity.Mathematics
         /// Checks if each component of the input is a power of two.
         /// </summary>
         /// <remarks>If a component of x is less than or equal to zero, then this function returns false in that component.</remarks>
-        /// <param name="x"><see cref="uint3"/> input</param>
-        /// <returns><see cref="bool3"/> where true in a component indicates the same component in the input was a power of two.</returns>
+        /// <param name="x"><see cref="Mathematics.uint3"/> input</param>
+        /// <returns><see cref="Mathematics.bool3"/> where true in a component indicates the same component in the input was a power of two.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool3 ispow2(uint3 x)
         {
@@ -519,8 +521,8 @@ namespace Unity.Mathematics
         /// Checks if each component of the input is a power of two.
         /// </summary>
         /// <remarks>If a component of x is less than or equal to zero, then this function returns false in that component.</remarks>
-        /// <param name="x"><see cref="uint4"/> input</param>
-        /// <returns><see cref="bool4"/> where true in a component indicates the same component in the input was a power of two.</returns>
+        /// <param name="x"><see cref="Mathematics.uint4"/> input</param>
+        /// <returns><see cref="Mathematics.bool4"/> where true in a component indicates the same component in the input was a power of two.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool4 ispow2(uint4 x)
         {
@@ -3648,7 +3650,7 @@ namespace Unity.Mathematics
         /// <remarks>
         /// Components of x must be greater than 0, otherwise the result for that component is undefined.
         /// </remarks>
-        /// <param name="x"><see cref="int2"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.int2"/> to be used as input.</param>
         /// <returns>Componentwise ceiling of the base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int2 ceillog2(int2 x)
@@ -3662,7 +3664,7 @@ namespace Unity.Mathematics
         /// <remarks>
         /// Components of x must be greater than 0, otherwise the result for that component is undefined.
         /// </remarks>
-        /// <param name="x"><see cref="int3"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.int3"/> to be used as input.</param>
         /// <returns>Componentwise ceiling of the base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int3 ceillog2(int3 x)
@@ -3676,7 +3678,7 @@ namespace Unity.Mathematics
         /// <remarks>
         /// Components of x must be greater than 0, otherwise the result for that component is undefined.
         /// </remarks>
-        /// <param name="x"><see cref="int4"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.int4"/> to be used as input.</param>
         /// <returns>Componentwise ceiling of the base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int4 ceillog2(int4 x)
@@ -3704,7 +3706,7 @@ namespace Unity.Mathematics
         /// <remarks>
         /// Components of x must be greater than 0, otherwise the result for that component is undefined.
         /// </remarks>
-        /// <param name="x"><see cref="uint2"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.uint2"/> to be used as input.</param>
         /// <returns>Componentwise ceiling of the base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int2 ceillog2(uint2 x)
@@ -3718,7 +3720,7 @@ namespace Unity.Mathematics
         /// <remarks>
         /// Components of x must be greater than 0, otherwise the result for that component is undefined.
         /// </remarks>
-        /// <param name="x"><see cref="uint3"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.uint3"/> to be used as input.</param>
         /// <returns>Componentwise ceiling of the base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int3 ceillog2(uint3 x)
@@ -3732,7 +3734,7 @@ namespace Unity.Mathematics
         /// <remarks>
         /// Components of x must be greater than 0, otherwise the result for that component is undefined.
         /// </remarks>
-        /// <param name="x"><see cref="uint4"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.uint4"/> to be used as input.</param>
         /// <returns>Componentwise ceiling of the base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int4 ceillog2(uint4 x)
@@ -3756,7 +3758,7 @@ namespace Unity.Mathematics
         /// Computes the componentwise floor of the base-2 logarithm of x.
         /// </summary>
         /// <remarks>Components of x must be greater than zero, otherwise the result of the component is undefined.</remarks>
-        /// <param name="x"><see cref="int2"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.int2"/> to be used as input.</param>
         /// <returns>Componentwise floor of base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int2 floorlog2(int2 x)
@@ -3768,7 +3770,7 @@ namespace Unity.Mathematics
         /// Computes the componentwise floor of the base-2 logarithm of x.
         /// </summary>
         /// <remarks>Components of x must be greater than zero, otherwise the result of the component is undefined.</remarks>
-        /// <param name="x"><see cref="int3"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.int3"/> to be used as input.</param>
         /// <returns>Componentwise floor of base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int3 floorlog2(int3 x)
@@ -3780,7 +3782,7 @@ namespace Unity.Mathematics
         /// Computes the componentwise floor of the base-2 logarithm of x.
         /// </summary>
         /// <remarks>Components of x must be greater than zero, otherwise the result of the component is undefined.</remarks>
-        /// <param name="x"><see cref="int4"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.int4"/> to be used as input.</param>
         /// <returns>Componentwise floor of base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int4 floorlog2(int4 x)
@@ -3804,7 +3806,7 @@ namespace Unity.Mathematics
         /// Computes the componentwise floor of the base-2 logarithm of x.
         /// </summary>
         /// <remarks>Components of x must be greater than zero, otherwise the result of the component is undefined.</remarks>
-        /// <param name="x"><see cref="uint2"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.uint2"/> to be used as input.</param>
         /// <returns>Componentwise floor of base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int2 floorlog2(uint2 x)
@@ -3816,7 +3818,7 @@ namespace Unity.Mathematics
         /// Computes the componentwise floor of the base-2 logarithm of x.
         /// </summary>
         /// <remarks>Components of x must be greater than zero, otherwise the result of the component is undefined.</remarks>
-        /// <param name="x"><see cref="uint3"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.uint3"/> to be used as input.</param>
         /// <returns>Componentwise floor of base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int3 floorlog2(uint3 x)
@@ -3828,7 +3830,7 @@ namespace Unity.Mathematics
         /// Computes the componentwise floor of the base-2 logarithm of x.
         /// </summary>
         /// <remarks>Components of x must be greater than zero, otherwise the result of the component is undefined.</remarks>
-        /// <param name="x"><see cref="uint4"/> to be used as input.</param>
+        /// <param name="x"><see cref="Mathematics.uint4"/> to be used as input.</param>
         /// <returns>Componentwise floor of base-2 logarithm of x.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int4 floorlog2(uint4 x)